kafka/sales-forecast/consumer/requirements.txt
kafka-python-ng==2.2.3
numpy
pandas
python-dotenv
requests
supabase
psycopg[binary]
